test_reporting/jupyter-notebooks/requirements.txt
asttokens==2.4.1
azure-core==1.32.0
azure-identity==1.19.0
azure-kusto-data==4.6.1
backcall==0.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
comm==0.2.2
contourpy==1.1.1
cryptography==43.0.3
cycler==0.12.1
debugpy==1.8.8
decorator==5.1.1
et-xmlfile==2.0.0
executing==2.1.0
fonttools==4.54.1
idna==3.10
ijson==3.3.0
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
ipywidgets==8.1.6
jedi==0.19.2
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyterlab-widgets==3.0.14
kiwisolver==1.4.7
matplotlib==3.7.5
matplotlib-inline==0.1.7
msal==1.31.0
msal-extensions==1.2.0
nest-asyncio==1.6.0
numpy==1.24.4
openpyxl==3.1.5
packaging==24.2
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.3.6
portalocker==2.10.1
prompt-toolkit==3.0.48
psutil==6.1.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycparser==2.22
pygments==2.18.0
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
widgetsnbextension==4.0.14
zipp==3.20.2
